requirements.txt
annotated-types==0.7.0
anyio==4.9.0
atproto==0.0.59
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.0
cryptography==44.0.2
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.7.1
dnspython==2.7.0
dotenv==0.9.9
exceptiongroup==1.2.2
fonttools==4.56.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
kiwisolver==1.4.7
libipld==3.0.1
matplotlib==3.9.4
numpy==2.0.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
tinycss2==1.4.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
webencodings==0.5.1
websockets==13.1
zipp==3.21.0
